requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
oauthlib==3.0.1
PySocks==1.6.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
